Requirement: #1139
Another partial commit.  Extensive refactoring of the compression implementation.  We now have 1 type of compression rule, which is applied at the StudyProcess time.  This rule can insert into the FilesystemQueue multiple entries for compression. Specific types of compression are configured at this time, and an XML field has been added to the FilesystemQueue page to set the compression parameters.  Either a lossless or lossy compress entry is placed into the FilesystemQueue table at this time.  There still remains a lossless and lossy service lock processor. We now just copy the XML details of the compress into the WorkQueue's XML field.  There is now a single WorkQueue type for handling any compression.  It takes the XML data in the WorkQueue entry and determines what compression type to apply, and also uses that XML to set the parameters for the compression.  This changeset adds the ability to set the compression ratio for JPEG 2000 Lossy and set the quality factor for JPEG Baseline/Extended.  This is done in the rules, and per rule.  As part of the changeset, did some mods to the Jpeg 2000 codec and rechecked it in.  The compression ratio codec parameter is now a float instead of an int, for better settings of the ratio.

This could use a bit more polish and testing.

+++ Dicom/Codec/Jpeg2000/DicomJpeg2000Codec.cpp
@@ -135,12 +135,12 @@ void DicomJpeg2000Codec::Encode(DicomUncompressedPixelData^ oldPixelData, DicomC
 			for (; r < jparams->RateLevels->Length; r++) {
 				if (jparams->RateLevels[r] > jparams->Rate) {
 					eparams.tcp_numlayers++;
-					eparams.tcp_rates[r] = (float)jparams->RateLevels[r];
+					eparams.tcp_rates[r] = jparams->RateLevels[r];
 				} else
 					break;
 			}
 			eparams.tcp_numlayers++;
-			eparams.tcp_rates[r] = (float)jparams->Rate;
+			eparams.tcp_rates[r] = jparams->Rate;
 		} else {
 			eparams.tcp_rates[0] = 0;
 			eparams.tcp_numlayers = 1;
@@ -251,7 +251,7 @@ void DicomJpeg2000Codec::Encode(DicomUncompressedPixelData^ oldPixelData, DicomC
 				
 				double oldSize = oldPixelData->UncompressedFrameSize;
 				double newSize = newPixelData->GetCompressedFrameSize(0);
-				newPixelData->LossyImageCompressionRatio = oldSize / newSize;
+				newPixelData->LossyImageCompressionRatio = (float) (oldSize / newSize);
 			}
 		}
 		finally {
